Fix signed/unsigned comparison mismatches.

+++ src/hal/user_comps/sendkeys.c
@@ -271,7 +271,7 @@ int main(int argc, char* argv[]) {
             }
             if (*hal->keycode != param->oldcode) {
                 /* Key press, report the event, send key release, and report again*/
-                if ((*hal->keycode & 0x3F) > param->num_events) continue;
+                if ((int)(*hal->keycode & 0x3F) > param->num_events) continue;
                 if (hal->event[*hal->keycode & 0x3F] == 0) continue;
                 if ((*hal->keycode & 0xC0) == 0xC0){ // keydown
                     emit(param->fd, EV_KEY, hal->event[*hal->keycode & 0x3F], 1);
